chore: move logic to `CustomDataset` class
