+++ src/distilabel/dataset.py
@@ -16,15 +16,24 @@ import warnings
 from dataclasses import dataclass, field
 from os import PathLike
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Dict, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Union
 
 from datasets import Dataset
 
 from distilabel.utils.dataset import load_task_from_disk, save_task_to_disk
 from distilabel.utils.imports import _ARGILLA_AVAILABLE
 
+if _ARGILLA_AVAILABLE:
+    from argilla.client.feedback.integrations.sentencetransformers import (
+        SentenceTransformersExtractor,
+    )
+
+
 if TYPE_CHECKING:
-    from argilla import FeedbackDataset
+    from argilla import FeedbackDataset, FeedbackRecord
+    from argilla.client.feedback.integrations.sentencetransformers import (
+        SentenceTransformersExtractor,
+    )
 
     from distilabel.tasks.base import Task
 
@@ -37,9 +46,17 @@ class CustomDataset(Dataset):
 
     task: Union["Task", None] = None
 
-    def to_argilla(self) -> "FeedbackDataset":
+    def to_argilla(
+        self,
+        vector_strategy: Union[bool, "SentenceTransformersExtractor"] = True,
+    ) -> "FeedbackDataset":
         """Converts the dataset to an Argilla `FeedbackDataset` instance, based on the
         task defined in the dataset as part of `Pipeline.generate`.
+
+        Args:
+            vector_strategy (Union[bool, SentenceTransformersExtractor]): the strategy to be used for
+                adding vectors to the dataset. If `True`, the default `SentenceTransformersExtractor`
+                will be used with the `TaylorAI/bge-micro-2` model. If `False`, no vectors will be added to the dataset.
 
         Raises:
             ImportError: if the argilla library is not installed.
@@ -93,7 +110,39 @@ class CustomDataset(Dataset):
                     UserWarning,
                     stacklevel=2,
                 )
+        rg_dataset = self.add_vectors_to_argilla_dataset(
+            dataset=rg_dataset, vector_strategy=vector_strategy
+        )
+
         return rg_dataset
+
+    def add_vectors_to_argilla_dataset(
+        self,
+        dataset: Union["FeedbackRecord", List["FeedbackRecord"], "FeedbackDataset"],
+        vector_strategy: Union[bool, "SentenceTransformersExtractor"],
+    ) -> Union["FeedbackRecord", List["FeedbackRecord"], "FeedbackDataset"]:
+        if _ARGILLA_AVAILABLE and vector_strategy:
+            try:
+                if isinstance(vector_strategy, SentenceTransformersExtractor):
+                    ste: SentenceTransformersExtractor = vector_strategy
+                elif vector_strategy:
+                    ste = SentenceTransformersExtractor()
+
+                dataset = ste.update_dataset(dataset=dataset)
+            except Exception as e:
+                warnings.warn(
+                    f"An error occurred while adding vectors to the dataset: {e}",
+                    stacklevel=2,
+                )
+
+        elif not _ARGILLA_AVAILABLE and vector_strategy:
+            warnings.warn(
+                "An error occurred while adding vectors to the dataset: "
+                "The `argilla`/`sentence-transformers` packages are not installed or the installed version is not compatible with the"
+                " required version. If you want to add vectors to your dataset, please run `pip install 'distilabel[vectors]'`.",
+                stacklevel=2,
+            )
+        return dataset
 
     def save_to_disk(self, dataset_path: PathLike, **kwargs: Any) -> None:
         """Saves the datataset to disk, also saving the task.
