+++ src/distilabel/tasks/text_generation/base.py
@@ -32,9 +32,6 @@ if _ARGILLA_AVAILABLE:
 
 if TYPE_CHECKING:
     from argilla import FeedbackDataset, FeedbackRecord
-    from argilla.client.feedback.integrations.sentencetransformers import (
-        SentenceTransformersExtractor,
-    )
 
 
 @dataclass
@@ -158,7 +155,6 @@ class TextGenerationTask(Task):
         self,
         dataset_row: Dict[str, Any],
         generations_column: Optional[str] = "generations",
-        vector_strategy: Union[bool, "SentenceTransformersExtractor"] = True,
     ) -> "FeedbackDataset":
         # First we infer the fields from the input_args_names, but we could also
         # create those manually instead using `rg.TextField(...)`
@@ -206,15 +202,11 @@ class TextGenerationTask(Task):
                 )
         # Then we just return the `FeedbackDataset` with the fields, questions, and metadata properties
         # defined above.
-        dataset = rg.FeedbackDataset(
+        return rg.FeedbackDataset(
             fields=fields,
             questions=questions,
             metadata_properties=metadata_properties,  # Note that these are always optional
         )
-        dataset = Task.add_vectors_to_argilla_dataset(
-            dataset=dataset, vector_strategy=vector_strategy
-        )
-        return dataset
 
     def to_argilla_record(self, dataset_row: Dict[str, Any]) -> "FeedbackRecord":
         """Converts a dataset row to an Argilla `FeedbackRecord`."""
