+++ src/distilabel/tasks/text_generation/self_instruct.py
@@ -15,9 +15,9 @@
 import re
 import warnings
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
-from distilabel.tasks.base import Task, get_template
+from distilabel.tasks.base import get_template
 from distilabel.tasks.prompt import Prompt
 from distilabel.tasks.text_generation.base import TextGenerationTask
 from distilabel.utils.argilla import (
@@ -31,9 +31,6 @@ if _ARGILLA_AVAILABLE:
 
 if TYPE_CHECKING:
     from argilla import FeedbackDataset, FeedbackRecord
-    from argilla.client.feedback.integrations.sentencetransformers import (
-        SentenceTransformersExtractor,
-    )
 
 _SELF_INSTRUCT_TEMPLATE = get_template("self-instruct.jinja2")
 
@@ -106,11 +103,7 @@ class SelfInstructTask(TextGenerationTask):
         pattern = re.compile(r"\d+\.\s*(.*?)\n")
         return {"instructions": pattern.findall(output)}
 
-    def to_argilla_dataset(
-        self,
-        dataset_row: Dict[str, Any],
-        vector_strategy: Union[bool, "SentenceTransformersExtractor"] = True,
-    ) -> "FeedbackDataset":
+    def to_argilla_dataset(self, dataset_row: Dict[str, Any]) -> "FeedbackDataset":
         # First we infer the fields from the input_args_names, but we could also
         # create those manually instead using `rg.TextField(...)`
         fields = infer_fields_from_dataset_row(
@@ -156,17 +149,12 @@ class SelfInstructTask(TextGenerationTask):
         )  # type: ignore
         # Then we just return the `FeedbackDataset` with the fields, questions, and metadata properties
         # defined above.
-        dataset = rg.FeedbackDataset(
+
+        return rg.FeedbackDataset(
             fields=fields,
             questions=questions,  # type: ignore
             metadata_properties=metadata_properties,  # Note that these are always optional
         )
-        dataset: (
-            FeedbackRecord | List[FeedbackRecord] | FeedbackDataset
-        ) = Task.add_vectors_to_argilla_dataset(
-            dataset=dataset, vector_strategy=vector_strategy
-        )
-        return dataset
 
     def to_argilla_record(
         self,
